+++ gaas-ui/src/util/sanitize-inputs.js
